api-geek/requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.4.2
cryptography==2.1.3
Django==1.11.5
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-swagger==2.1.2
djangorestframework==3.7.0
Fabric==1.14.0
fake==0.8
Faker==0.8.6
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.4
openapi-codec==1.3.2
paramiko==2.3.1
psycopg2==2.7.3.1
pyasn1==0.3.7
pycparser==2.18
PyNaCl==1.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
rest-condition==1.0.3
simplejson==3.11.1
six==1.11.0
text-unidecode==1.1
uritemplate==3.0.0
urllib3==1.22
